Move 'BugReporterAdapter' to '/src/app' to uncouple lib from 'BugReporter'
